Add "Upgrading" status to the condition CR
A pod readiness probe is used in Openshift to communicate the readiness of the service, but in OLM it also means "upgrade is done" for operators. This is sometimes creates conflicts when the upgrade is a very long process, longer than the readiness probe timeout. This conflicts may cause OLM to kill and restart the operator pod, in some cases.

PR #42 introduces the condition CR - a way for the operator to communicate with OLM. This PR suggests a new condtion, `Upgrading`, to try to solve the above conflict, by moving the "upgrade is done" detection from the readiness probe to the new conditon.