Set deprecated and new property names statically
Coding the GRPC server to handle on-the-fly property renaming
and duplication would mean a compatibility matrix would need
to exist between the GRPC server and its clients.

To avoid that problem, we'll rely on build-time and
validation-time tooling to help index maintainers build
declarative configs that work with both old and new clients.

Signed-off-by: Joe Lanford <joe.lanford@gmail.com>