declarative config: document how bundle objects are referenced and stored for back-compat
The `olm.bundle.object` property can be used to include bundle objects
in an index. It behaves like a union type, where one of `ref` or `data`
must be set. If `ref` is set, `opm` will load data from the referenced
file path, relative to the directory in which the `olm.bundle` blob is
found. If `data` is set, `opm` will base64 decode the `data` value.

For backwards-compatibility reasons, the `olm.bundle.object` properties
are necessary for `opm` to serve OLM's packageserver component with
`CSV`-derived metadata from a declarative config-based index. The API
service exposed by packageserver is used by various on-cluster tools,
such as the OpenShift console and the `kubectl operator` plugin. Index
maintainers that desire for their indexes to support these components
should validate that CSV objects are present in the index for each
channel head of each package.

Signed-off-by: Joe Lanford <joe.lanford@gmail.com>