Update olm-toggle-copied-csvs enhancement
This commit removed the reference of copied RBAC from the
olm-toggle-copied-csvs enhancement. When an operator is
installed in `AllNamespace` mode its permissions are
elevated to cluster level and are not installed in each
namespace. Additionally, the ClusterRoles and ClusterRoleBindings
are created with an "owner label" pointing to the original CSV,
which is an OLM construct that allows it to clean up resources
related to the CSV when it is deleted.

This commit also a spelling mistake, updates the design to
create a single event notifying users that the Copied CSVs
feature has been disabled, and removes mention of the
command line option to disable copied csvs.